+++ src/widgets/world/RenderWindow.cpp
@@ -192,6 +192,7 @@ RenderWindow::RenderWindow(QWidget* parent)
     : QWidget(parent),
       p(std::make_unique<Private>()) {
   p->graph = std::make_shared<engine::graph::GraphRenderEngine>(nullptr);
+  p->graph->setExecutionTraceEnabled(true);
 
   auto grid = new QGridLayout(this);
   p->settingsWidget = new RenderSettingsWidget(this);
@@ -243,6 +244,8 @@ void RenderWindow::render() {
   p->time.restart();
 
   p->settingsWidget->setBusy(true);
+  p->graphInspector->setExecutionTrace(nullptr);
+  p->graphInspector->clearExecutionState();
 
   p->renderWidget->resize(p->settingsWidget->resolution());
   p->renderWidget->setBufferSize(p->settingsWidget->resolution());
@@ -271,6 +274,8 @@ void RenderWindow::stop() {
 void RenderWindow::finished() {
   p->settingsWidget->setBusy(false);
   p->busy = false;
+  p->graphInspector->setExecutionTrace(
+    p->graph->lastExecutionTraceForPlan(p->graphInspector->effectivePlan()));
 }
 
 void RenderWindow::setScene(::Scene* scene) {